Remove ecp policy checker
Removes the ECP policy eligibility checker as we're removing new claims
for this policy we don't need to be able to check their eligibility.
