+++ spec/models/policies/targeted_retention_incentive_payments/policy_eligibility_checker_spec.rb
@@ -209,126 +209,4 @@ RSpec.describe Policies::TargetedRetentionIncentivePayments::PolicyEligibilityCh
       end
     end
   end
-
-  describe "#eligible_now?" do
-    before { create(:journey_configuration, :targeted_retention_incentive_payments) }
-
-    subject { policy_eligibility_checker.eligible_now? }
-
-    context "eligible now" do
-      let(:answers) {
-        build(
-          :targeted_retention_incentive_payments_answers,
-          :targeted_retention_incentive_eligible
-        )
-      }
-
-      it { is_expected.to be true }
-    end
-
-    context "eligible later" do
-      let(:answers) {
-        build(
-          :targeted_retention_incentive_payments_answers,
-          :targeted_retention_incentive_eligible_later
-        )
-      }
-
-      it { is_expected.to be false }
-    end
-  end
-
-  describe "#eligible_later?" do
-    before do
-      create(
-        :journey_configuration,
-        :targeted_retention_incentive_payments,
-        current_academic_year: AcademicYear.new(2023)
-      )
-    end
-
-    subject { policy_eligibility_checker.eligible_later? }
-
-    context "eligible now" do
-      let(:answers) {
-        build(
-          :targeted_retention_incentive_payments_answers,
-          :targeted_retention_incentive_eligible
-        )
-      }
-
-      it { is_expected.to be false }
-    end
-
-    context "eligible later" do
-      let(:answers) {
-        build(
-          :targeted_retention_incentive_payments_answers,
-          :targeted_retention_incentive_eligible_later
-        )
-      }
-
-      it { is_expected.to be true }
-    end
-  end
-
-  describe "#status" do
-    before do
-      create(
-        :journey_configuration,
-        :targeted_retention_incentive_payments,
-        current_academic_year: AcademicYear.new(2023)
-      )
-    end
-
-    subject { policy_eligibility_checker.status }
-
-    it_behaves_like "eligibility_status", :targeted_retention_incentive_payments
-
-    context "ECP-only ITT subject" do
-      let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :targeted_retention_incentive_ineligible_itt_subject) }
-
-      it { is_expected.to eq(:ineligible) }
-    end
-
-    context "ITT subject or degree subject" do
-      context "good ITT subject and no degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :no_relevant_degree) }
-
-        it { is_expected.to eq(:eligible_now) }
-      end
-
-      context "bad ITT subject but have a degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :targeted_retention_incentive_ineligible_itt_subject, :relevant_degree) }
-
-        it { is_expected.to eq(:eligible_now) }
-      end
-
-      context "bad ITT subject and no degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :targeted_retention_incentive_ineligible_itt_subject, :no_relevant_degree) }
-
-        it { is_expected.to eq(:ineligible) }
-      end
-    end
-
-    context "trainee teacher" do
-      context "good ITT subject and no degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :trainee_teacher, :no_relevant_degree) }
-
-        it { is_expected.to eq(:eligible_later) }
-      end
-
-      context "bad ITT subject but have a degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :trainee_teacher, :targeted_retention_incentive_ineligible_itt_subject, :relevant_degree) }
-
-        it { is_expected.to eq(:eligible_later) }
-      end
-
-      context "bad ITT subject and no degree" do
-        let(:answers) { build(:targeted_retention_incentive_payments_answers, :targeted_retention_incentive_eligible, :trainee_teacher, :targeted_retention_incentive_ineligible_itt_subject, :no_relevant_degree) }
-
-        it { is_expected.to eq(:ineligible) }
-      end
-    end
-  end
 end
